agents/coder/requirements.txt
aiohttp
prometheus-api-client
docker
git+https://github.com/QuantGeekDev/docker-mcp.git
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-distro
